Don't store resource data in project_c
The project_c class used to store a wav_c object for the project's WAV
audio; but this commit moves that responsibility to the GUI's WAV player,
and leaves project_c to simply provide the name of the WAV file.

This is required, since should project_c store resources, it should
logically also store the textual transcription. However, the transcription
is also stored in the GUI's text editor widget, which has no model/view
pattern; leading to the need to manually synchronize data between the GUI
and the project_c object.

Instead, I think it's better to have project_c not as a container of data
but as a provider of metadata, which in that sense simply ties together
the project's attributes.