Add loading projects via dragging folder onto main window
This commit adds the tarpaulin, a widget that covers the main window
while no project is loaded in. In doing so, it hides all the GUI editing
controls that aren't needed while no project is loaded in.

Additionally, the tarp watches for drag/drop events from the user. When a
folder containing a project's files is dragged onto the window, the tarp
is pulled back, the GUI controls enabled, and the project's data loaded
in for editing.