Add creating new project by dragging media file to main window
The program now allows the user to create a new transcription project by
dragging a media file onto the main window (while the tarp is on).

FFmpeg is then used via a system() call to extract a ranssi-compatible
WAV file from the file the user dragged in; after which a new project is
created with that WAV file, and opened for editing.