Text editor: Strip block's formatting when block is active
Removes formatting (like color) from a block's text while the text cursor
is within that block. The formatting is reinstated when the cursor leaves
the block.

This is a bit of a kludge to prevent formatting from 'smudging' when new
text is inserted within that formatting's span. For instance, inserting
new text inside a boldface span will cause the new text to be in bold as
well, without a convenient way of making it normal.